:recycle: rename project name
