+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'praise-jail'
+rootProject.name = 'praise-push'
