+++ src/test/java/com/praise/praisejail/PraisePushApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PraiseJailApplicationTests {
+class PraisePushApplicationTests {
 
     @Test
     void contextLoads() {
